% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get PMCID from DOI}
\usage{
get_id(doi)
}
\arguments{
\item{doi}{a vector or dataframe column containing digital object identifier(s)}
}
\value{
a dataframe containing the original DOI(s) and the new IDs (PMID and PMCID)
}
\description{
Get PMCID from DOI
}
