% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_ft.R
\name{get_xml_ft}
\alias{get_xml_ft}
\title{Get XML from EuropePMC}
\usage{
get_xml_ft(path, pmcid)
}
\arguments{
\item{path}{a file path to the folder where XMLs should be saved}

\item{pmcid}{a vector or dataframe column containing PMCIDs}
}
\value{
no objects returned, XML files saved to path specified
}
\description{
Get XML from EuropePMC
}
